Refactor: MVC 패턴으로 구조 재구성 및 메서드 주석 추가
