+++ src/App.js
@@ -1,6 +1,6 @@
-const OutputView = require('./OutputView');
-const InputView = require('./InputView');
-const BridgeGame = require('./BridgeGame');
+const OutputView = require('./view/OutputView');
+const InputView = require('./view/InputView');
+const BridgeGame = require('./model/BridgeGame');
 
 class App {
   play() {
