+++ src/view/OutputView.js
@@ -1,13 +1,12 @@
 const MissionUtils = require('@woowacourse/mission-utils')
-const { MESSAGE } = require('./Messages');
+const { MESSAGE } = require('../Messages');
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
 const OutputView = {
   /**
    * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
-   * <p>
-   * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
+   * @param {BridgeGame} bridgeGame 현재 실행 중인 다리 게임
    */
   printMap(bridgeGame) {
     let [upHistory, downHistory] = bridgeGame.getUpDownHistory();
@@ -17,8 +16,7 @@ const OutputView = {
 
   /**
    * 게임의 최종 결과를 정해진 형식에 맞춰 출력한다.
-   * <p>
-   * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
+   * @param {BridgeGame} bridgeGame 현재 실행 중인 다리 게임
    */
   printResult(bridgeGame) {
     MissionUtils.Console.print(MESSAGE.GAME_RESULT);
@@ -27,10 +25,17 @@ const OutputView = {
     MissionUtils.Console.print(`${MESSAGE.GAME_TRY_COUNT}${bridgeGame.getTryCount()}`);
   },
 
+  /**
+  * 게임 시작 문구를 출력한다.
+  */
   printStart() {
     MissionUtils.Console.print(MESSAGE.START);
   },
 
+  /**
+   * 에러 메세지를 출력한다.
+   * @param {string} message 에러 메세지
+   */
   printError(message) {
     MissionUtils.Console.print(message);
   }
